refactor(core): split patterns/affinity into submodules
Split large affinity.rs file into module directory structure:
- mod.rs: Public API and exports
- classifier.rs: PatternAffinityClassifier implementation
- computation.rs: Context similarity computation
- types.rs: RelativeAffinity, EpisodeAssignmentGuard, RejectionReason
- tests.rs: Unit tests (fixed module_inception clippy lint)

File size compliance: Each file now <500 LOC.
